refactor sidebar layout and move
